+++ src/dialogs/ReplaceDialog.cpp
@@ -148,7 +148,7 @@ void CReplaceDialog::OnDestroy()
 	pApp->WriteProfileInt( REGKEY_SEARCH_DIALOG, "MatchCase", m_bMatchCase );
 	pApp->WriteProfileInt( REGKEY_SEARCH_DIALOG, "RegularExpression", m_bRegularExpression );
 
-	INT nCount = m_lstFindString.GetCount();
+	INT nCount = (INT)m_lstFindString.GetCount();
 	pApp->WriteProfileInt( REGKEY_SEARCH_DIALOG, "FindStringCount", nCount );
 
 	POSITION pos = m_lstFindString.GetHeadPosition();
@@ -158,7 +158,7 @@ void CReplaceDialog::OnDestroy()
 		pApp->WriteProfileString( REGKEY_SEARCH_DIALOG, szEntry, szFindString );
 	}
 
-	INT nCoun2 = m_lstReplaceString.GetCount();
+	INT nCoun2 = (INT)m_lstReplaceString.GetCount();
 	pApp->WriteProfileInt( REGKEY_SEARCH_DIALOG, "ReplaceStringCount", nCoun2 );
 
 	POSITION po2 = m_lstReplaceString.GetHeadPosition();
